Update create_TEIs.py
If metadata versions differ, do not exit. metadata version is not a reliable way of measuring if the metadata which matters for dummy data creation is ok. For example, if we change the OU assignment, this value changes but the script should be able to work without needing to call create_flat_file again