Remove future warning
Solving warnings:
One related to the use of pandas' concat function which is likely due to a change in how pandas will handle object-dtype columns that contain only boolean values in future versions. The warning suggests to explicitly cast these columns to bool dtype to avoid issues in future versions of pandas.
Another issue when new_row_df is created using pd.DataFrame([new_row], columns=df.columns), it's possible that the data types of the columns in new_row_df are not matching those in df, especially for boolean columns.