Update VLOOKUP formula to catch multiple UIDs
From Pablo: When you have 2 programs and the programStages are having the same name, the FF is assigning the wrong programStage. This is coming because when we export the PR metadata, we're creating a list which is including only "unique" programstages names.
For example:
You have
Program A - ProgramStage "Laboratory" (UID xxxxx)
Program B - ProgramStage "Laboratory" (UID yyyyy)
You export the PR and all PRs that are triggered for Program Stage "Laboratory" are getting the UID xxxxxx. Meaning that the PR for Program B are broken